requirements.txt
black==20.8b1
coverage==5.3.1
flake8==3.8.4
iniconfig==1.1.1
nose==1.3.7
numpy==1.19.5
py==1.10.0
pylint==2.6.0
pytesseract==0.3.7
pyzmq==22.0.2
opencv-python==4.5.1.48
pyserial==3.5

